requirements.txt
boto3
appdirs==1.4.3
attrs==19.3.0
aws-cdk.assets==1.19.0
aws-cdk.aws-apigateway==1.16.3
aws-cdk.aws-applicationautoscaling==1.19.0
aws-cdk.aws-autoscaling==1.19.0
aws-cdk.aws-autoscaling-common==1.19.0
aws-cdk.aws-autoscaling-hooktargets==1.19.0
aws-cdk.aws-certificatemanager==1.19.0
aws-cdk.aws-cloudformation==1.19.0
aws-cdk.aws-cloudfront==1.19.0
aws-cdk.aws-cloudwatch==1.16.3
aws-cdk.aws-codebuild==1.19.0
aws-cdk.aws-codecommit==1.19.0
aws-cdk.aws-codepipeline==1.19.0
aws-cdk.aws-ec2==1.16.3
aws-cdk.aws-ecr==1.19.0
aws-cdk.aws-ecr-assets==1.19.0
aws-cdk.aws-ecs==1.19.0
aws-cdk.aws-elasticbeanstalk==1.20.0
aws-cdk.aws-elasticloadbalancing==1.19.0
aws-cdk.aws-elasticloadbalancingv2==1.19.0
aws-cdk.aws-events==1.16.3
aws-cdk.aws-events-targets==1.19.0
aws-cdk.aws-iam==1.16.3
aws-cdk.aws-kms==1.19.0
aws-cdk.aws-lambda==1.16.3
aws-cdk.aws-logs==1.16.3
aws-cdk.aws-route53==1.19.0
aws-cdk.aws-route53-targets==1.19.0
aws-cdk.aws-s3==1.16.3
aws-cdk.aws-s3-assets==1.16.3
aws-cdk.aws-secretsmanager==1.19.0
aws-cdk.aws-servicediscovery==1.19.0
aws-cdk.aws-sns==1.19.0
aws-cdk.aws-sns-subscriptions==1.19.0
aws-cdk.aws-sqs==1.16.3
aws-cdk.aws-ssm==1.16.3
aws-cdk.aws-stepfunctions==1.19.0
aws-cdk.core==1.16.3
aws-cdk.cx-api==1.16.3
aws-cdk.region-info==1.16.3
awscli==1.16.283
black==19.10b0
boto==2.49.0
botocore==1.13.19
cattrs==0.9.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
docutils==0.15.2
idna==2.8
jmespath==0.9.4
jsii==0.20.5
pathspec==0.6.0
publication==0.0.3
pyasn1==0.4.7
python-dateutil==2.8.1
PyYAML==5.1.2
regex==2019.12.19
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.6
virtualenv==16.7.7
aws-cdk.aws-sagemaker==1.16.3
